Created validation for the remove stock form.

+++ SMS-2.0/RemoveStock.cs
@@ -19,14 +19,24 @@ namespace SMS_2._0
 
         private void RemoveStockBttn_Click(object sender, EventArgs e)
         {
-           Database database = new Database();
-            int newQuantity = database.GetQuantity(int.Parse(idTXTBOX.Text)) - int.Parse(quantityTXTBOX.Text);
-            database.RunQuery($"UPDATE stock SET quantity = {newQuantity} WHERE stock.id = {idTXTBOX.Text}");
-            if (database.GetQuantity(int.Parse(idTXTBOX.Text)) == 0)
+            if (int.Parse(quantityTXTBOX.Text) > 0 )
             {
-                database.RunQuery($"DELETE FROM `stock` WHERE `stock`.`id` = {int.Parse(idTXTBOX.Text)}");
+                Database database = new Database();
+                int newQuantity = database.GetQuantity(int.Parse(idTXTBOX.Text)) - int.Parse(quantityTXTBOX.Text);
+                database.RunQuery($"UPDATE stock SET quantity = {newQuantity} WHERE stock.id = {idTXTBOX.Text}");
+                if (database.GetQuantity(int.Parse(idTXTBOX.Text)) == 0)
+                {
+                    database.RunQuery($"DELETE FROM `stock` WHERE `stock`.`id` = {int.Parse(idTXTBOX.Text)}");
+                }
+                this.Close();
             }
-            this.Close();
+            else
+            {
+                MessageBox.Show("Quantity must be positive");
+                quantityTXTBOX.Text = "";
+                idTXTBOX.Text = "";
+            }
+          
         }
     }
 }
